Finish audio input request mode on Android.

+++ architecture/android/app/src/main/java/com/faust/FaustActivity.java
@@ -88,6 +88,7 @@ implements ActivityCompat.OnRequestPermissionsResultCallback {
     
     // For audio input request permission
     private static final int AUDIO_ECHO_REQUEST = 0;
+    private boolean permissionToRecordAccepted = false;
     
 	private SensorManager mSensorManager;
 	private int numberOfParameters;
@@ -115,9 +116,9 @@ implements ActivityCompat.OnRequestPermissionsResultCallback {
         boolean isImmersiveModeEnabled =
                 ((uiOptions | View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY) == uiOptions);
         if (isImmersiveModeEnabled) {
-            Log.i("FaustJava", "Turning immersive mode mode off.");
+            Log.i("FaustJava", "Turning immersive mode mode off");
         } else {
-            Log.i("FaustJava", "Turning immersive mode mode on.");
+            Log.i("FaustJava", "Turning immersive mode mode on");
         }
 
         // Navigation bar hiding:  Backwards compatible to ICS.
@@ -145,14 +146,13 @@ implements ActivityCompat.OnRequestPermissionsResultCallback {
         getWindow().getDecorView().setSystemUiVisibility(newUiOptions);
         //END_INCLUDE (set_ui_flags)
     }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
+    
+    private void createFaust() {
         
-        Log.d("FaustJava", "onCreate");
+        Log.d("FaustJava", "createFaust");
         sensorIntervalMs = bufferSize/sampleRate*1000;
         
-        if (dspFaust == null) {          
+        if (dspFaust == null) {
             // check if audio files were saved in internal storage
             String[] internalStorageList = getFilesDir().list();
             boolean fileWereCopied = false;
@@ -162,7 +162,7 @@ implements ActivityCompat.OnRequestPermissionsResultCallback {
                     break;
                 }
             }
-
+            
             // if audio files were not saved in internal storage, then transfer them
             if(!fileWereCopied) {
                 AssetManager assets = getAssets();
@@ -175,7 +175,7 @@ implements ActivityCompat.OnRequestPermissionsResultCallback {
                             in = assets.open(assetsList[i]);
                             File outFile = new File(getFilesDir().getPath(), assetsList[i]);
                             out = new FileOutputStream(outFile);
-
+                            
                             // copy content
                             byte[] buffer = new byte[1024];
                             int read;
@@ -189,12 +189,12 @@ implements ActivityCompat.OnRequestPermissionsResultCallback {
                             out = null;
                         }
                     }
-
+                    
                 } catch (IOException e) {
                     e.printStackTrace();
                 }
             }
-
+            
             WifiManager wifi = (WifiManager)getSystemService(Context.WIFI_SERVICE);
             if (wifi != null) {
                 WifiManager.MulticastLock lock = wifi.createMulticastLock("Log_Tag");
@@ -204,51 +204,43 @@ implements ActivityCompat.OnRequestPermissionsResultCallback {
             int oscPortNumber = 5510;
             while (!Osc.init(oscPortNumber)) oscPortNumber++;
             Log.d("FaustJava", "onCreate : OSC In Port " + oscPortNumber);
-
+            
             // Use machine buffer size and sample rate
             AudioManager audioManager = (AudioManager) this.getSystemService(Context.AUDIO_SERVICE);
             
             // Do not work on Android 4.xx
             /*
-            String rate = audioManager.getProperty(AudioManager.PROPERTY_OUTPUT_SAMPLE_RATE);
-            String size = audioManager.getProperty(AudioManager.PROPERTY_OUTPUT_FRAMES_PER_BUFFER);
-            Log.d("FaustJava", "Size :" + size + " Rate: " + rate);
-            dsp_faust.init(Integer.parseInt(rate), Integer.parseInt(size));
-            */
+             String rate = audioManager.getProperty(AudioManager.PROPERTY_OUTPUT_SAMPLE_RATE);
+             String size = audioManager.getProperty(AudioManager.PROPERTY_OUTPUT_FRAMES_PER_BUFFER);
+             Log.d("FaustJava", "Size :" + size + " Rate: " + rate);
+             dsp_faust.init(Integer.parseInt(rate), Integer.parseInt(size));
+             */
             
-            // For audio input request permission
+            // TODO: sr and buffer length should change in function of the device for best latency perfs
+            dspFaust = new DspFaust(sampleRate, bufferSize);
             
-            if (!isRecordPermissionGranted()){
-                requestRecordPermission();
-                return;
-            } else {
-                // TODO: sr and buffer length should change in function of the device for best latency perfs
-                dspFaust = new DspFaust(sampleRate, bufferSize);
-            }
-
             Osc.startListening();
         }
-
+        
         fBuildUI = (dspFaust.getScreenColor() < 0);
         if (!fBuildUI) {
             requestWindowFeature(Window.FEATURE_NO_TITLE);
-			getWindow().addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+            getWindow().addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
         }
-        super.onCreate(savedInstanceState);
         if (fBuildUI) {
             setContentView(R.layout.main);
         } else {
             fMonoView = new MonochromeView(getApplicationContext());
             setContentView(fMonoView);
         }
-
+        
         numberOfParameters = dspFaust.getParamsCount();
-
+        
         Log.d("FaustJava", "onCreate : numberOfParameters " + numberOfParameters);
-
+        
         parametersInfo.init(numberOfParameters);
         SharedPreferences settings = getSharedPreferences("savedParameters", 0);
-
+        
         if (fBuildUI) {
             LinearLayout mainGroup = (LinearLayout) findViewById(R.id.the_layout);
             HorizontalScrollView horizontalScroll = (HorizontalScrollView) findViewById(R.id.horizontalScroll);
@@ -259,8 +251,22 @@ implements ActivityCompat.OnRequestPermissionsResultCallback {
             Log.d("FaustJava", "Don't create User Interface");
             toggleHideyBar();
         }
-
+        
         mSensorManager = (SensorManager) getSystemService(Context.SENSOR_SERVICE);
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        
+        Log.d("FaustJava", "onCreate");
+        super.onCreate(savedInstanceState);
+        
+        if (Build.VERSION.SDK_INT >= 23 && !isRecordPermissionGranted()){
+            requestRecordPermission();
+        } else {
+            permissionToRecordAccepted = true;
+            createFaust();
+        }
     }
 
     private void updatecolor(double x, double y, double z) {
@@ -389,26 +395,30 @@ implements ActivityCompat.OnRequestPermissionsResultCallback {
     @Override
    	protected void onPause() {
         Log.d("FaustJava", "onPause");
-   		mSensorManager.unregisterListener(mSensorListener);
-   		super.onPause();
+        if (permissionToRecordAccepted) {
+            mSensorManager.unregisterListener(mSensorListener);
+        }
+        super.onPause();
    	}
 
     @Override
     protected void onResume() {
         Log.d("FaustJava", "onResume");
-		super.onResume();
-        mSensorManager.registerListener(mSensorListener, mSensorManager.getDefaultSensor(
-	    		Sensor.TYPE_ACCELEROMETER), SensorManager.SENSOR_DELAY_FASTEST);
+        super.onResume();
+        if (permissionToRecordAccepted) {
+            if (!isChangingConfigurations()) {
+                dspFaust.start();
+            }
+            ui.updateUIstate();
+            mSensorManager.registerListener(mSensorListener, mSensorManager.getDefaultSensor(
+                    Sensor.TYPE_ACCELEROMETER), SensorManager.SENSOR_DELAY_FASTEST);
+        }
    }
 
     @Override
     protected void onStart() {
         Log.d("FaustJava", "onStart");
         super.onStart();
-        if (!isChangingConfigurations()) {
-            dspFaust.start();
-        }
-        ui.updateUIstate();
     }
 
     @Override
@@ -425,7 +435,7 @@ implements ActivityCompat.OnRequestPermissionsResultCallback {
         // correctly work
         /*
         if (!isChangingConfigurations()) {
-            dsp_faust.stop();
+            dspFaust.stop();
         }
         */
     }
@@ -433,15 +443,17 @@ implements ActivityCompat.OnRequestPermissionsResultCallback {
     @Override
     public void onDestroy() {
         Log.d("FaustJava", "onDestroy");
-    	// only stops audio when the user press the return button (and not when the screen is rotated)
-    	if (!isChangingConfigurations()) {
-            Osc.stopListening();
-            dspFaust.stop(); // TODO: not sure if needed
-    		dspFaust.delete();
-            dspFaust = null;
+        if (permissionToRecordAccepted) {
+            // only stops audio when the user press the return button (and not when the screen is rotated)
+            if (!isChangingConfigurations()) {
+                Osc.stopListening();
+                dspFaust.stop(); // TODO: not sure if needed
+                dspFaust.delete();
+                dspFaust = null;
+            }
+            SharedPreferences settings = getSharedPreferences("savedParameters", 0);
+            parametersInfo.saveParameters(settings);
         }
-        SharedPreferences settings = getSharedPreferences("savedParameters", 0);
-        parametersInfo.saveParameters(settings);
         super.onDestroy();
     }
 
@@ -456,22 +468,13 @@ implements ActivityCompat.OnRequestPermissionsResultCallback {
 
         if (grantResults.length != 1 ||
             grantResults[0] != PackageManager.PERMISSION_GRANTED) {
-
-            // User denied the permission, without this we cannot record audio
-            // Show a toast and update the status accordingly
-            //statusText.setText(R.string.status_record_audio_denied);
-            //statusText.setText("status_record_audio_denied");
-            Toast.makeText(getApplicationContext(),
-                            //getString(R.string.need_record_audio_permission),
-                            "need_record_audio_permission",
-                            Toast.LENGTH_SHORT)
-            .show();
+            finish();
         } else {
-            // Permission was granted, start echoing
-            //toggleEcho();
+            // Permission was granted
+            permissionToRecordAccepted = true;
+            createFaust();
         }
     }
-
 
     // For audio input request permission
     private boolean isRecordPermissionGranted() {
